Add files via upload
Task 1:
Unit Convertor Application

Description:
This application helps users convert units from one form to another. It takes input from the user and converts it to the desired unit, providing the result. For example, it converts values from centimeters to meters or grams to kilograms and vice versa.

Task 2:
To Do App

Description:
This application, as the name suggests, holds a To-Do list for users. It stores their schedules, upcoming meetings, and events. Users can also jot down important notes. To ensure security, a login page is provided before accessing the application.

This app encompasses a login page, sign-up page, and a logout system. Users have an area to write their tasks, events, or important notes. It is built using Java and XML in Android Studio, ensuring a user-friendly interface. To securely store users' data, SQLite(Room) is employed, allowing users to even permanently delete their data.

Task 3:
Calculator

Description:
This application is a straightforward calculator designed for those who are just starting to learn Android development and Java coding. It takes input values and the desired operation from the users. After receiving the input, it displays the results on the screen. This application is simple and doesn't require the use of any specific packages.

Task 4:
Quiz Application

This application presents users with questions and provides multiple-choice answers. Users select their chosen answer, and on-screen feedback indicates if the answer is correct. At the end of the quiz, the application displays the final score to the user.

Task 5:
Stop Watch

Description:
This application functions as a digital stopwatch, mirroring the capabilities of a traditional handheld timepiece. It measures the time elapsed between activation and deactivation. The application includes three buttons: start, stop, and hold.